// frontend/src/redux-api/server-sent-events/event-types.ts
export enum ServerSentEventType {
  // Dokumenter under arbeid
  DOCUMENT_FINISHED = 'DOCUMENT_FINISHED',
  DOCUMENTS_ADDED = 'DOCUMENTS_ADDED',
  DOCUMENTS_REMOVED = 'DOCUMENTS_REMOVED',
  DOCUMENTS_CHANGED = 'DOCUMENTS_CHANGED',

  /** Smartdokumenter */
  SMART_DOCUMENT_LANGUAGE = 'SMART_DOCUMENT_LANGUAGE',
  // New version created
  SMART_DOCUMENT_VERSIONED = 'SMART_DOCUMENT_VERSIONED',
  // New comment or reply
  SMART_DOCUMENT_COMMENT_ADDED = 'SMART_DOCUMENT_COMMENT_ADDED',
  // Comment or reply removed
  SMART_DOCUMENT_COMMENT_REMOVED = 'SMART_DOCUMENT_COMMENT_REMOVED',
  // Comment or reply changed
  SMART_DOCUMENT_COMMENT_CHANGED = 'SMART_DOCUMENT_COMMENT_CHANGED',

  // Journalførte dokumenter
  JOURNALFOERT_DOCUMENT_MODIFIED = 'JOURNALFOERT_DOCUMENT_MODIFIED',
  JOURNALPOST_ADDED = 'JOURNALPOST_ADDED',
  INCLUDED_DOCUMENTS_ADDED = 'INCLUDED_DOCUMENTS_ADDED',
  INCLUDED_DOCUMENTS_REMOVED = 'INCLUDED_DOCUMENTS_REMOVED',
  INCLUDED_DOCUMENTS_CLEARED = 'INCLUDED_DOCUMENTS_CLEARED',

  // Behandlingsdetaljer
  TILDELING = 'TILDELING',
  ROL = 'ROL',
  MEDUNDERSKRIVER = 'MEDUNDERSKRIVER',
  KLAGER = 'KLAGER',
  FULLMEKTIG = 'FULLMEKTIG',
  GOSYSOPPGAVE = 'GOSYSOPPGAVE',
  /** Added or changed */
  MESSAGE = 'MESSAGE',
  UTFALL = 'UTFALL',
  EXTRA_UTFALL = 'EXTRA_UTFALL',
  REGISTRERINGSHJEMLER = 'REGISTRERINGSHJEMLER',
  INNSENDINGSHJEMLER = 'INNSENDINGSHJEMLER',
  MOTTATT_VEDTAKSINSTANS = 'MOTTATT_VEDTAKSINSTANS',
  SATT_PAA_VENT = 'SATT_PAA_VENT',
  FERDIGSTILT = 'FERDIGSTILT',
  FEILREGISTRERING = 'FEILREGISTRERING',
  TILBAKEKREVING = 'TILBAKEKREVING',
  VARSLET_FRIST = 'VARSLET_FRIST',
}